scripts: add script to extract all closed PRs that need backporting
All closed PRs with label "stable/needs-backport" are extracted
and upstream commit ids are listed. This helps to avoid doing
this manually by going through the PRs and is less error prone
and time consuming. The script is just an fugly hack (improvements
welcome!), but good enough to get the job done.

One strict requirement for this is that "stable/needs-backport"
is properly maintained on GH. Meaning, after the backport work
was merged into a stable branch, this needs to be removed from
the PR.

At some point we could automate that further with 'cherry-pick'
script. ;-) This could work for conflict-free backports at least.

Usage and output example:

$ ./check-stable
Checking for a valid github API token: OK
PRs for stable backporting: 2

 * PR: 2274 -- Misc cleanup and possible flake fix (@scanf)
   Merge with 5 commit(s) merged at: Wed, 06 Dec 2017 22:22:51 +0100!
                                                v (start)
     * e4211427b53a0031c000324699f61cf2d312275e |
     * 9c6d05499cee768a4b052904ed3559fdf5287b43 |
     * 08c5ec720f40edd3402e73d4723baa5f80ecf750 |
     * eebc7a976ef2ce475270b29c79bfe3d4485d32b6 |
     * 320a67c5cb298c3aec2e4c7e97d7f6eaaa387408 |
                                                v (end)

 * PR: 2271 -- Fix `cilium bpf policy list` to print l4 ports (@joestringer)
   Merge with 6 commit(s) merged at: Thu, 07 Dec 2017 22:32:14 +0100!
                                                v (start)
     * 91313637c17d15cef1e54b941ab20ec4dd91a8ad |
     * 31c65a9b64c4c2fd032bd5499236b0c9099cd932 |
     * 6a24efde730892c4489b1daeb4e152461ae9fcda |
     * 79497c1f35bddff4f6a4e18f57f892616ca6111e |
     * c8642c41dd180ff8eed22623f38f07961fb3b1e4 |
     * f247e93f42be2664e0741139c5bf32125fa94660 |
                                                v (end)

This can then be used for cherry-picking season:

$ ./cherry-pick e4211427b53a0031c000324699f61cf2d312275e
$ ./cherry-pick 9c6d05499cee768a4b052904ed3559fdf5287b43
[...]

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>